LPS-55590 Copy ExportUsersAction, no logic changes
